+++ CHANGELOG.md
@@ -64,3 +64,4 @@
 * Define `ifelse` function for `Boolean` values
 * Change `Stack` and `Nonempty` definitions, temporarily remove `filter`
 * Define `<**>`, `<***>`, `<****>` methods for `Applicative` to compact expressions
+* Change `Graph` definition